+++ packages/generator/src/resources/emoticons.ts
@@ -22,7 +22,9 @@
  *    variants are automatically generated.
  */
 
-export const emoticons = {
+import { EmoticonMap } from '../types';
+
+export const emoticons: EmoticonMap = {
   // HAPPY
   // 🙂 slightly smiling face
   '1F642': ':)',
@@ -37,7 +39,7 @@ export const emoticons = {
   // 😬 grimacing face
   '1F62C': '8D',
   // 😇 smiling face with halo
-  '1F607': 'o:)',
+  '1F607': ['o:)', 'O:)'],
   // 😈 smiling face with horns
   '1F608': '>:)',
 
@@ -47,13 +49,13 @@ export const emoticons = {
   // 🤣 rolling on the floor laughing
   '1F923': ":'D",
   // 😆 smiling face with open mouth & closed eyes
-  '1F606': 'xD',
+  '1F606': ['xD', 'XD'],
   // 😛 face with stuck-out tongue
-  '1F61B': ':p',
+  '1F61B': [':p', ':P'],
   // 😜 face with stuck-out tongue & winking eye
-  '1F61C': ';p',
+  '1F61C': [';p', ';P'],
   // 😝 face with stuck-out tongue & closed eyes
-  '1F61D': 'xp',
+  '1F61D': ['xp', 'xP', 'XP'],
   // 😏 smirking face
   '1F60F': ':j',
 
@@ -63,7 +65,7 @@ export const emoticons = {
   // 😚 kissing face with closed eyes
   '1F61A': ':*',
   // 😘 face blowing a kiss
-  '1F618': ':x',
+  '1F618': [':x', ':X'],
 
   // NEUTRAL
   // 😐️ neutral face
@@ -75,15 +77,15 @@ export const emoticons = {
   // 😓 face with cold sweat
   '1F613': ':<',
   // 🤐 zipper-mouth face
-  '1F910': ':z',
+  '1F910': [':z', ':Z'],
 
   // SAD
   // 🙁 frowning face
   '2639': ':(',
   // 😧 anguished face
-  '1F627': ':s',
+  '1F627': [':s', ':S'],
   // 😖 confounded face
-  '1F616': 'x(',
+  '1F616': ['x(', 'X('],
   // 😢 crying face
   '1F622': ":'(",
   // 😭 loudly crying face
@@ -97,17 +99,17 @@ export const emoticons = {
   // 👿 angry face with horns
   '1F47F': '>:(',
   // 😫 tired face
-  '1F62B': ':c',
+  '1F62B': [':c', ':C'],
   // woozy face
   '1F974': ':&',
 
   // MISC
   // 🤔 thinking face
-  '1F914': ':l',
+  '1F914': [':l', ':L'],
   // 😲 astonished face
-  '1F632': ':o',
+  '1F632': [':o', ':O'],
   // 😵 dizzy face
-  '1F635': 'xo',
+  '1F635': ['xo', 'XO'],
   // 🤬 face with symbols over mouth
   '1F92C': ':@',
   // 🤢 nauseated face
@@ -135,5 +137,5 @@ export const emoticons = {
   // 💔 broken heart
   '1F494': '</3',
   // 🤘 sign of the horns
-  '1F918': '\\m/',
+  '1F918': ['\\m/', '\\M/'],
 };